+++ CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ## [Unreleased]
 
+## [1.14.0] - 2026-03-11
+
+### Added
+
+- **Top Projects & Top Branches Bar Charts** - New horizontal bar chart panels in the Dashboard displayed side-by-side:
+  - Top Projects chart showing where you've spent the most time
+  - Top Branches chart showing most active branches by time spent
+  - Color-coded bars with time axis labels
+  - Responsive layout that adapts to panel width
+
+### Fixed
+
+- Pomodoro skip break now works correctly from idle state
+
 ## [1.13.0] - 2026-03-09
 
 ### Added
@@ -247,7 +261,8 @@
 
 - Support IntelliJ Platform 2024.3.5
 
-[Unreleased]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.13.0...HEAD
+[Unreleased]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.14.0...HEAD
+[1.14.0]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.13.0...v1.14.0
 [1.13.0]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.12.0...v1.13.0
 [1.12.0]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.11.0...v1.12.0
 [1.11.0]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.10.0...v1.11.0